@@ -1,20 +1,25 @@
-import React, { Component } from "react";
+import { useState } from "react";
 import Header from "./Header";
 import Order from "./Order";
 import Inventory from "./Inventory";
 
-class App extends Component {
-  render() {
-    return( 
-      <div className="catch-of-the-day">
-        <div className="menu">
-          <Header />
-        </div>
-        <Order />
-        <Inventory />
+function App() {
+  const [fishes, setFishes] = useState([]);
+  const [order, setOrder] = useState("");
+
+  const addFish = fish => {
+    setFishes([ ...fishes, fish ]);
+  };
+
+  return (
+    <div className="catch-of-the-day">
+      <div className="menu">
+        <Header tagline="Fresh Seafood Market" />
       </div>
-    );
-  }
+      <Order />
+      <Inventory addFish={addFish} />
+    </div>
+  );
 }
 
 export default App;
