@@ -1,9 +1,7 @@
 import React from "react";
 
-class Header extends React.Component {
-  render() {
-    return (
-      <header className="top">
+const Header = (props) => (
+  <header className="top">
         <h1>Catch
           <span className="ofThe">
             <span className="of">Of</span>
@@ -12,11 +10,9 @@ class Header extends React.Component {
           Day
         </h1>
         <h3 className="tagline">
-          <span>Fresh Daily</span>
+          <span>{props.tagline}</span>
         </h3>
       </header>
-    );
-  }
-}
+);
 
 export default Header;
