@@ -1,13 +1,12 @@
-import React from "react";
+import AddFishForm from "./AddFishForm";
 
-class Inventory extends React.Component {
-  render() {
-    return (
-      <div className="inventory">
-        Inventory!
-      </div>
-    );
-  }
+function Inventory({addFish}) {
+  return (
+    <div className="inventory">
+      <h2>Inventory</h2>
+      <AddFishForm addFish={addFish} />
+    </div>
+  );
 }
 
 export default Inventory;
