@@ -1,13 +1,10 @@
-import React from "react";
 
-class Order extends React.Component {
-  render() {
-    return (
-      <div className="order">
-        Order!
-      </div>
-    );
-  }
+function Order() {
+  return (
+    <div className="order">
+      Order!
+    </div>
+  );
 }
 
 export default Order;
